Silence -Wshadow and qsizetype format warnings
Coalesce THUMTOO_DEBUG / BILTOO_THUMTOO_DEBUG getenv into one scope; print
ladder byte length as long long.

+++ src/imageloader.cpp
@@ -916,8 +916,8 @@ QImage loadThumbnail(const QString &path, int maxEdge)
             }
             // Bytes exist but neither VIPS nor Qt could decode — do not call
             // schedulePixels again (that re-emits ladderReady and spins the pool).
-            qWarning("ImageLoader: ladder payload undecodable for %s (maxEdge=%d, %d bytes)",
-                     qPrintable(path), maxEdge, ladder.size());
+            qWarning("ImageLoader: ladder payload undecodable for %s (maxEdge=%d, %lld bytes)",
+                     qPrintable(path), maxEdge, static_cast<long long>(ladder.size()));
             return {};
         }
         // True miss: ask thumtoo to build the level in the background for next time.
